Move change_email_for_login into lib/ret.ex
